increase z-index to get over sites that use very high numbers
For example, http://passpo.jp/page/pageProfile/ and then click on one of the crew to open the profile popup; it has a z-index of 99999.

I chose 999999999 which is not the maximum number that works but should be clear to anyone reading the code that it's an arbitrary extremely large number. (It's the largest number made only of 9s that would work.) Reference: http://softwareas.com/whats-the-maximum-z-index/